fix d et i for int min

+++ print_functions.c
@@ -54,10 +54,10 @@ int print_int(va_list args)
 	int temp;
 	int digits;
 	unsigned int argument = va_arg(args, unsigned int);
-	int num = argument;
+	unsigned int num = argument;
 	int count = 0;
 
-	if (num < 0)
+	if ((int)num < 0)
 	{
 		count += _putchar('-');
 		num = -num;
